+++ VK-Challenge/VK-Challenge/Core/API/Model/VKAPIArrayResponse.swift
@@ -1,0 +1,13 @@
+//
+//  VKAPIArrayResponse.swift
+//  VK-Challenge
+//
+//  Created by Evgeniy on 09.11.18.
+//  Copyright © 2018 Evgeniy. All rights reserved.
+//
+
+import Foundation
+
+struct VKAPIArrayResponse<T: Decodable>: Decodable {
+    let response: [T]
+}
