+++ VK-Challenge/VK-Challenge/Core/API/VKAPI+Type.swift
@@ -1,0 +1,18 @@
+//
+//  VKAPIMethod.swift
+//  VK-Challenge
+//
+//  Created by Evgeniy on 09.11.18.
+//  Copyright © 2018 Evgeniy. All rights reserved.
+//
+
+import Foundation
+
+protocol VKAPIMethod {
+    var value: String { get }
+}
+
+protocol VKAPIField {
+    var value: String { get }
+}
+
